Added favicon and upload is now working

+++ src/app/views/admin/admin.component.ts
@@ -5,24 +5,34 @@ import { AdminService } from './service/admin.service';
 import { ConfirmationService } from 'primeng/api';
 import { MessageService } from 'primeng/api';
 
-
-
 @Component({
   templateUrl: './admin.component.html',
-  styleUrls: ['./admin.component.scss']
+  styleUrls: ['./admin.component.scss'],
 })
 export class AdminComponent implements OnInit {
-  ads$: Observable<IAd[]> = this.adminService.ads$
-  constructor(private adminService: AdminService) { }
+  ads$: Observable<IAd[]> = this.adminService.ads$;
+  constructor(
+    private adminService: AdminService,
+    private confirmationService: ConfirmationService
+  ) {}
 
-  deleteAd(id:string) {
-    this.adminService.deleteAd(id).subscribe();
-  }
-  updateStatus(id: string) {
-    this.adminService.updateStatus(id).subscribe();
-  }
   ngOnInit(): void {
     this.adminService.getAd().subscribe();
   }
-
+  deleteAd(id: string) {
+    this.confirmationService.confirm({
+      message: 'Da li želite da obrišete ovaj oglas?',
+      accept: () => {
+        this.adminService.deleteAd(id).subscribe();
+      },
+    });
+  }
+  updateAd(id: string) {
+    this.confirmationService.confirm({
+      message: 'Da li želite da promenite status oglasa?',
+      accept: () => {
+        this.adminService.updateStatus(id).subscribe();
+      },
+    });
+  }
 }
